Add vanilla CSS support to Tabs (#510)
