+++ data/primitives/components/tabs/1.0.1.mdx
@@ -16,9 +16,7 @@ aria: https://www.w3.org/WAI/ARIA/apg/patterns/tabpanel
   <TabsDemo />
 </HeroContainer>
 
-```jsx hero template=Tabs
-
-```
+<HeroCodeBlock folder="Tabs" />
 
 <Highlights
   features={[
